+++ src/app/documents/page.tsx
@@ -4,49 +4,12 @@ import DocumentsAccordionItem from "~/components/accordion/documents-accordion-i
 import SectionTitle from "~/components/layout/section-title";
 import SectionWrapper from "~/components/layout/section-wrapper";
 import { Accordion } from "~/components/ui/accordion";
-import { prisma } from "~/server/db";
+import { getDocumentGroupsWithDocuments } from "./get-document-groups-with-documents";
 
 export const metadata: Metadata = {
   title: "Dokument",
   description: "Här kan du hitta alla Z-teknologsektionens dokument",
 };
-
-export const getDocumentGroupsWithDocuments = async (): Promise<
-  {
-    Document: {
-      isPDF: boolean;
-      title: string;
-      url: string;
-    }[];
-    extraText: string;
-    id: string;
-    name: string;
-  }[]
-> =>
-  await prisma.documentGroup.findMany({
-    where: {
-      Document: {
-        some: {
-          id: {
-            not: undefined,
-          },
-        },
-      },
-    },
-    select: {
-      id: true,
-      name: true,
-      extraText: true,
-      Document: {
-        select: {
-          id: true,
-          isPDF: true,
-          title: true,
-          url: true,
-        },
-      },
-    },
-  });
 
 const DocumentsPage: FC = async () => {
   const documentGroups = await getDocumentGroupsWithDocuments();
